+++ demo-app/src/main/java/org/openmined/syft/demo/service/TrainingTask.kt
@@ -1,4 +1,4 @@
-package org.openmined.syft.demo
+package org.openmined.syft.demo.service
 
 import androidx.work.ListenableWorker.Result
 import androidx.work.workDataOf
