+++ demo-app/src/main/res/layout/activity_mnist.xml
@@ -102,18 +102,39 @@
                         android:textAlignment="viewStart"
                         app:contentPadding="2dp" />
                 </ScrollView>
-
-                <Button
-                    android:id="@+id/button_first"
-                    style="@android:style/Widget.DeviceDefault.Button"
+                <LinearLayout
                     android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:background="@drawable/button_solid"
-                    android:onClick="@{() -> viewModel.startCycle()}"
-                    android:padding="10dp"
-                    android:text="@string/start"
-                    android:textColor="@color/white"
-                    android:textStyle="bold" />
+                    android:layout_height="match_parent"
+                    android:gravity="center_vertical"
+                    android:orientation="vertical">
+
+                    <Button
+                        android:id="@+id/button_first"
+                        style="@android:style/Widget.DeviceDefault.Button"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:background="@drawable/button_solid"
+                        android:onClick="@{() -> viewModel.startCycle()}"
+                        android:padding="10dp"
+                        android:text="@string/start_foreground"
+                        android:textColor="@color/white"
+                        android:textSize="12sp"
+                        android:textStyle="bold" />
+
+                    <Button
+                        android:id="@+id/button_second"
+                        style="@android:style/Widget.DeviceDefault.Button"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_marginTop="20dp"
+                        android:background="@drawable/button_solid"
+                        android:padding="10dp"
+                        android:text="@string/start_background"
+                        android:textColor="@color/white"
+                        android:textSize="12sp"
+                        android:textStyle="bold" />
+                </LinearLayout>
+
             </LinearLayout>
 
         </androidx.constraintlayout.widget.ConstraintLayout>
