+++ demo-app/src/main/res/values/strings.xml
@@ -3,7 +3,8 @@
     <string name="log_message_hint">Syft will tell you what happens in this area</string>
     <string name="title_activity_base_u_r_l">FullscreenActivity</string>
     <string name="submit">Submit</string>
-    <string name="start">Start Cycle</string>
+    <string name="start_background">Start Background</string>
+    <string name="start_foreground">Start Foreground</string>
     <string name="enter_url">Enter PyGrid server URL:</string>
     <string name="error_url">url is not valid</string>
 </resources>
